Move stories to be closer to source
