+++ src/components/Privacy.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import PrivacyComponent from '../../components/Privacy';
+import PrivacyComponent from './Privacy';
 
 export default {
   title: 'TOS/Privacy',
